Generator writes HTML output files
Guide generator writes the result of the Markdown parsing to an .html
file using the title as the filename. The file is written to the output
folder, within a subfolder for the current language.

The parser now also uses extra settings for a more customised output.
Details on these extras can be found here:
https://github.com/trentm/python-markdown2/wiki/Extras

Issues that will arise later:
- Handling multiple languages in one output document.
- Handling files that were left over from the previous generation