Restructured the parser matches text and calls function
We now use a list (for order) of regex's and the function to call
if matched. The parser now uses a Parser object to access data easily.
The parser now looks at each text section (text broken on page breaks)
and replaces the matched regexed text, with the result of the function
call when passing the regex match through.

TODO: Will need to store both the list of code to subsitute in another
file for ease of access.