Restructured file hierarchy and settings locations
This commit completes a lot of small steps. These will be tested
with the guide content once they been changed to using dashes in
filenames.

This commit:
- Other_files folder is renamed to files
- Output-generator is renamed to generator
- Moves files users would use to the top level, these
  are generateguide.py and guide-settings.py
- Moved constant values into a settings
  file generator-settings.conf
- Removed some uneffecient code
- Added some docstrings

The order of processing text (chapters then appendices)
needs to be stored and used better.

There is probably issues with this version of the generator
however will some testing of new guide content, the
generator will be better structured for the future.