Modifications to section heading structure
Changes made to the way the structure of a section is stored.
Heading nodes now store a list 'children' of pointers to
sub-heading nodes.

Handles case where a heading has been missed (i.e a
level 4 heading following a level 2 heading) by creating
blank intermediate heading nodes in strucutre, and
logging the error.

Implemented __repr__ method for section object which
returns a representation of the heading structure of
the section, to aid with debugging.