Fixed SCSS import errors
The generator script now passes the SCSS
compiler a list of all subfolders in the
SCSS folder to search for files when
importing.

This could possibly cause issues in the
future (same file names), but the
alternative is the have all imports link
from the SCSS folder root, but this
would require editing Foundation scripts
on every update. Will need to see how
this works in the future.