Restructured log messages to include line number and section
When a regex function now needs to create
a log message (example: for an error), the
function creates a log message for this
regex. After all parsing, all log messages
are found in the original text and written
to the log file.

A big thank you to @jordangriffiths01 for
his hard work and help in making this change
happen.